+++ src/components/sections/elements/GeographicExtentMap.vue
@@ -1,12 +1,12 @@
 <script setup lang="ts">
 import { onMounted, type PropType } from 'vue'
 
-import type { WellKnownExtent } from '@/types/app'
+import type { GeographicExtent } from '@/types/iso'
 import { loadCssTheme, initExtentMap } from '@/lib/esriNoTest'
 
 const props = defineProps({
-  wke: {
-    type: Object as PropType<WellKnownExtent>,
+  extent: {
+    type: Object as PropType<GeographicExtent>,
     required: true,
   },
 })
@@ -15,11 +15,10 @@ const container = 'geographic-extent-map'
 
 onMounted(() => {
   loadCssTheme()
-  initExtentMap(container, props.wke)
+  initExtentMap(container, props.extent)
 })
 </script>
 
 <template>
   <div :id="container" class="h-96 w-full border border-sky-500"></div>
 </template>
-@/lib/esriNoTest